+++ CHANGELOG.md
@@ -14,6 +14,11 @@
 - generating `authors.txt`, `manufacturers.txt`, and `mpns.txt` for improving static hosting (e.g. github, github pages)
 - `jwtScopesPrefix` flag to set default prefix for scopes authentication
 - support for storing and searching/filtering TMs with MPN variations/placeholders
+- higher-level attachments import/fetch/delete/list in CLI
+- REST API: GET `/authors/{authorName}/.attachments/`, GET `/manufacturers/{authorName}/{manufacturerName}/.attachments/`, GET `/thing-models/.tmName/{tmName}/.attachments`, and GET `/thing-models/{tmID}/.attachments` to list attachments, which belong to the corresponding level (author/manufacturer/tmName/tmID, respectively)
+- REST API: GET `/authors/{authorName}/.attachments/{attachmentFileName}` and GET `/manufacturers/{authorName}/{manufacturerName}/.attachments/{attachmentFileName}` higher-level attachments fetch
+- REST API: POST `/authors/{authorName}/.attachments/{attachmentFileName}` and POST `/manufacturers/{authorName}/{manufacturerName}/.attachments/{attachmentFileName}` higher-level attachments import
+- REST API: DELETE `/authors/{authorName}/.attachments/{attachmentFileName}` and DELETE `/manufacturers/{authorName}/{manufacturerName}/.attachments/{attachmentFileName}` higher-level attachments delete
 - Added `filter.changedSince` parameter to REST API `/inventory` listing and to CLI
 
 ### Changed